creating api to get parsed json data and autofill on the
