+++ package.json
@@ -25,6 +25,7 @@
     "react-multi-carousel": "^2.8.4",
     "react-router-dom": "^6.14.2",
     "react-scripts": "5.0.1",
+    "req": "^0.1.4",
     "require": "^2.4.20",
     "stream": "^0.0.2",
     "styled-components": "^6.0.4",
