+++ server/server.js
@@ -2,14 +2,13 @@ const express = require('express');
 const multer = require('multer');
 const path = require('path');
 const cors = require('cors');
-const execa = require('execa'); // Import execa module
-
+const execa = require('execa');
+const fs = require('fs'); // Import the fs module
 const app = express();
+const {generateParsedFilename} = require('./utils');
 
-// Enable CORS
 app.use(cors());
 
-// Configure storage using Multer
 const storage = multer.diskStorage({
   destination: function (req, file, cb) {
     cb(null, path.join(__dirname, '../storage/inputresume'));
@@ -21,30 +20,53 @@ const storage = multer.diskStorage({
 
 const upload = multer({ storage });
 
-// Handle the file upload.
 app.post('/upload', upload.single('resume'), async (req, res) => {
+  const uploadedFileName = req.file.originalname;
+  const parsedFileName = uploadedFileName.replace(/\.[^.]+$/, ''); // Remove file extension
+
   try {
-    const uploadedFileName = req.file.originalname;
     const pythonScriptPath = 'src/models/parser-resume/ractor.py';
-    
-    // Execute Python script
-    const { stdout, stderr } = await execa('python', [pythonScriptPath, uploadedFileName]);
+    await execa('python', [pythonScriptPath, uploadedFileName]);
 
-    // Respond with a success message and the Python script's output
-    res.json({
-      message: 'Resume uploaded and parsed successfully!',
-      filename: uploadedFileName,
-      pythonOutput: stdout,
-    });
+    console.log('Python script executed successfully!');
 
-    console.log('Python script executed successfully:', stdout);
+    const jsonFilePath = path.join(__dirname, '../storage/output', `${parsedFileName}.json`);
+    const jsonData = fs.readFileSync(jsonFilePath, 'utf-8');
+    const resumeData = JSON.parse(jsonData);
+
+    const extractedInfo = {
+      email: resumeData.email,
+      phone: resumeData.phone,
+      linkedin: resumeData.linkedin,
+      github: resumeData.github,
+    };
+
+    res.write('Resume parsed successfully!'); // Send notification
+    res.end(); // End the response
+
+    console.log('Resume parsed successfully!');
   } catch (error) {
-    console.error('Error:', error.message);
-    res.status(500).json({ error: 'An error occurred.' });
+    console.error('Error executing Python script:', error);
+    res.status(500).json({ error: 'An error occurred while parsing the resume.' });
   }
 });
 
-// Default route to handle root URL
+app.get('/output/:filename', (req, res) => {
+  const parsedFileName = req.params.filename.replace(/\.[^.]+$/, ''); // Remove file extension
+  const jsonFilePath = path.join(__dirname, '../storage/output', `${parsedFileName}_parsed.json`);
+
+  if (fs.existsSync(jsonFilePath)) {
+    const jsonData = fs.readFileSync(jsonFilePath, 'utf-8');
+    const resumeData = JSON.parse(jsonData);
+    
+    res.write('Parsed resume data:\n');
+    res.write(JSON.stringify(resumeData, null, 2)); // Write JSON data
+    res.end(); // End the response
+  } else {
+    res.status(404).send('Parsed JSON file not found.');
+  }
+});
+
 app.get('/', (req, res) => {
   res.send('Server is running. Use POST request to /upload for file upload.');
 });
