--- storage/inputresume/Alex Dubinchyk.html
@@ -1,143 +1,0 @@
-<div id="readme" class="boxed-group flush clearfix announce instapaper_body md">
-	<h3>
-		<span class="octicon octicon-book"></span>
-		README.md
-	</h3>
-
-	<article class="markdown-body entry-content" itemprop="mainContentOfPage"><p><em>Alex Dubinchyk</em><br>
-		<em>Pleasant Hill, CA</em><br>
-		<em><a href="mailto:alexs.dbk@gmail.com">alexs.dbk@gmail.com</a></em><br>
-		<em>Status:Green Card Holder</em>
-		<br><br></p>
-
-		<h5>
-			<a id="user-content-objective"
-			   class="anchor"
-			   href="#objective"
-			   aria-hidden="true"><span class="octicon octicon-link"></span></a>OBJECTIVE</h5>
-
-		<p>Seeking a challenging position to use my software Web development and process optimization skills.</p>
-
-		<h5>
-			<a id="user-content-summary"
-			   class="anchor"
-			   href="#summary"
-			   aria-hidden="true"><span class="octicon octicon-link"></span></a>SUMMARY</h5>
-
-		<p>&nbsp;&nbsp;I worked on a wide range of products including building advanced dynamic multi language web sites,
-			internal and external API's, well as creating new internal workflows. My goal is to work in a passionate team,
-			that loves their work and the products they are creating together, supporting, mentoring, optimizing workflows and
-			creating high quality software.
-			I'm energenic, solution oriented team-player, constantly learning and growing as a team, and bringing high spirits
-			along with me.</p>
-
-		<h5>
-			<a id="user-content-technology"
-			   class="anchor"
-			   href="#technology"
-			   aria-hidden="true"><span class="octicon octicon-link"></span></a>Technology</h5>
-
-		<ul class="task-list">
-			<li> Server side PHP programming, REST, OPP, MVC, Yii framework.</li>
-			<li> SQL Database programming: MySQL, SQL, MSSQL.</li>
-			<li> Client-side programming: JavaScript, AJAX, jQuery.</li>
-			<li> Smarty Template Engine, HTML5, CSS3.</li>
-			<li> Webserver installation and configuration: Apache, Nginx, IIS.</li>
-			<li> Source control: SVN/Subversion, GIT.</li>
-			<li> Platforms: Linux, Mac, Windows.</li>
-		</ul>
-
-		<h5>
-			<a id="user-content-experience"
-			   class="anchor"
-			   href="#experience"
-			   aria-hidden="true"><span class="octicon octicon-link"></span></a>EXPERIENCE</h5>
-
-		<p><strong>Actuate <a href="http://www.actuate.com/">http://www.actuate.com/</a></strong><br>
-			<strong>Full stack PHP Developer</strong><br>
-			<strong>San Manteo, CA. November 2014 - current</strong></p>
-
-		<ul class="task-list">
-			<li>Implemented web service(API, MVC, Php)</li>
-			<li>Introduced Git to the team</li>
-		</ul>
-
-		<p><strong>BloomSky <a href="http://www.bloomsky.com">http://www.bloomsky.com</a></strong><br>
-			<strong>Backend Developer</strong><br>
-			<strong>Sunnyvale, CA. October 2014 - November 2014</strong></p>
-
-		<ul class="task-list">
-			<li>Integrated APIs (PHP, Python, Django, Nginx)</li>
-			<li>Set up PHPUnit and functional testing</li>
-			<li>Deploy merge DB script (MSSQL&gt;MySQL)</li>
-		</ul>
-
-		<p><strong>Rozumsoft LLC, / Telecontact LLC
-			<a href="http://www.telecontact.ru/">http://www.telecontact.ru/</a></strong><br>
-			<strong>Full stack PHP Developer</strong><br>
-			<strong>Belarus, Minsk. February 2012 - September 2014</strong><br><br>
-			&nbsp;&nbsp;<strong>Programming modules of dynamically building statistics for quality control assessment
-				project.</strong></p>
-
-		<ul class="task-list">
-			<li> Designed and developed project quality control assessment that estimated effective work of operators in
-				callcenters from different regions. Includes modules separation mapping based on more 20 users roles(RBAC),
-				online editors(logs, statistical formulas, projects rules and etc) for more 10k clients.
-			</li>
-			<li> Finalization coding script of internal protection algorithm authorization and validation.</li>
-			<li> Programming API service for quality control assessment. Interact with user interface(AJAX) with fast load up
-				JSON data, audio files and extract large data in excel.
-			</li>
-			<li> API data exchange integrated with data parse, merge, view in table linkage, send emails. Support more 100
-				source, near 1000 onlineusers, more 10 servers.
-			</li>
-			<li> Codes cross-browsers users interfaces in project quality control assessment, using Javascript, jQuery, JSON,
-				Bootstrap.
-			</li>
-			<li> Developed JavaScript audio player with individual custom design, hardware acceleration, deceleration and the
-				order to play audio files.
-			</li>
-		</ul>
-
-		<p>&nbsp;&nbsp;<strong>Designed and developed the company website (<a href="http://www.rozumsoft.com/">http://www.rozumsoft.com/</a>).</strong>
-		</p>
-
-		<ul class="task-list">
-			<li> Implemented 3 domain zones(ru/by/com) algorithm.</li>
-			<li> Codes contents editor for 3 languages</li>
-			<li> Fixed and support custom seo map logic.</li>
-		</ul>
-
-		<p>&nbsp;&nbsp;<strong>Developed scripts products to callcenters operators</strong></p>
-
-		<ul class="task-list">
-			<li> Development of сomplex reports and statistical summaries by Cisco data telephony.</li>
-			<li> Redesigned and reimplemented projects using MVC approach and strong OOP design</li>
-			<li> Designed and conversion of scripts database, extensive SQL query optimization.</li>
-		</ul>
-
-		<p><strong>Real Estate Agency Assistant heals LLC, Full stack PHP Developer</strong><br>
-			<strong>Belarus, Minsk, <a href="http://www.a-h.by">http://www.a-h.by</a>; May 2011 - January 2012</strong></p>
-
-		<ul class="task-list">
-			<li>Dynamic website design and programming using PHP, MySQL, HTML, CSS. Setup and administration of web servers
-				and server software.
-			</li>
-			<li>Business consulting of securing/ planning project.</li>
-			<li>Development to online marketing, search engine placement and promotion (<a href="http://www.mogu.by">http://www.mogu.by</a>;
-				<a href="http://www.a-h.by">http://www.a-h.by</a>).
-			</li>
-		</ul>
-
-		<h5>
-			<a id="user-content-education"
-			   class="anchor"
-			   href="#education"
-			   aria-hidden="true"><span class="octicon octicon-link"></span></a>EDUCATION</h5>
-
-		<ul class="task-list">
-			<li>Belarusian University of Informatics and Radioelectronics,</li>
-			<li>BS in Modeling and computer-aided design of radioelectronics devices.</li>
-		</ul>
-	</article>
-</div>
