--- storage/inputresume/arman.txt
@@ -1,37 +1,0 @@
-Arman
-
-Contact Information:
-
-Address: 123 Somewhere Road, Niceville, FL
-Phone: (555) 888-086483
-Email: arman@gmail.com
-Objective:
-An experienced MBA with a passion for digital marketing seeking a role as a Social Media Marketing Manager at XYZ Inc. to develop and implement successful marketing campaigns, and provide thought leadership to the team.
-
-Summary of Skills:
-
-Three years of specialized experience in 'Python', 'Java', 'SQL'
-Managed social media accounts for four businesses with yearly sales over $500,000
-Certified Digital Marketer 2016, Udacity
-Received Employee of the Year Award at ABC Company 2018 and 2019
-Work Experience:
-
-Full Stack Developer
-ABC Company, Nashville, TN, 2017-2020
-
-Built a 15k Facebook following and 20k Instagram following for a new startup, increasing leads by 160%
-Developed compelling CTAs for website and social media copy, producing a 60% increase in sales
-Formulated Instagram strategy and collaborated with influencers in our market to promote top-selling products, resulting in $200,000 net sales in three months
-Web Developer
-123 Company, Austin, TX, 2014-2017
-
-Team leader for the design and development of a new mobile application
-Conducted A/B testing on paid Facebook and Instagram campaigns, resulting in a boost in conversion from 70% to 90%
-Implemented an innovative SEO strategy to increase organic website traffic by 200% in five months
-Education:
-
-University of Michigan, 2006-2010
-
-B.tech, programmer
-References:
-Available upon request.
