--- storage/inputresume/hasley.txt
@@ -1,25 +1,0 @@
-def extract_content(paragraphs, patterns_dict):
-    '''
-    Extract headings and content based on the provided patterns.
-    '''
-    extracted_data = {}
-    current_heading = None
-    
-    for paragraph in paragraphs:
-        for heading, keywords in patterns_dict["titles"].items():
-            for keyword in keywords:
-                if keyword in paragraph.lower():
-                    current_heading = heading
-                    break
-        
-        if current_heading:
-            # Store the paragraph as content under the current heading
-            if current_heading not in extracted_data:
-                extracted_data[current_heading] = paragraph
-            else:
-                extracted_data[current_heading] += '\n' + paragraph
-        
-        current_heading = None  # Reset the current_heading
-    
-    return extracted_data
-
