--- storage/inputresume/resume.txt
@@ -1,80 +1,0 @@
-Alex Dubinchyk
-Pleasant Hill, CA
-alexs.dbk@gmail.com
-Status:Green Card Holder
-
-OBJECTIVE
-
-Seeking a challenging position to use my software Web development and process optimization skills.
-SUMMARY
-
-  I worked on a wide range of products including building advanced dynamic multi language web sites, internal and external API's, well as creating new internal workflows. My goal is to work in a passionate team, that loves their work and the products they are creating together, supporting, mentoring, optimizing workflows and creating high quality software. I'm energenic, solution oriented team-player, constantly learning and growing as a team, and bringing high spirits along with me.
-Technology
-
-    Server side PHP programming, REST, OPP, MVC, Yii framework.
-    SQL Database programming: MySQL, SQL, MSSQL.
-    Client-side programming: JavaScript, AJAX, jQuery.
-    Smarty Template Engine, HTML5, CSS3.
-    Webserver installation and configuration: Apache, Nginx, IIS.
-    Source control: SVN/Subversion, GIT.
-    Platforms: Linux, Mac, Windows.
-
-EXPERIENCE
-
-Actuate http://www.actuate.com/
-Full stack PHP Developer
-San Manteo, CA. November 2014 - current
-
-    Implemented web service(API, MVC, Php)
-    Introduced Git to the team
-
-BloomSky http://www.bloomsky.com
-Backend Developer
-Sunnyvale, CA. October 2014 - November 2014
-
-    Integrated APIs (PHP, Python, Django, Nginx)
-    Set up PHPUnit and functional testing
-    Deploy merge DB script (MSSQL>MySQL)
-
-Rozumsoft LLC, / Telecontact LLC http://www.telecontact.ru/
-Full stack PHP Developer
-Belarus, Minsk. February 2012 - September 2014
-
-  Programming modules of dynamically building statistics for quality control assessment project.
-
-    Designed and developed project quality control assessment that estimated effective work of operators in callcenters from different regions. Includes modules separation mapping based on more 20 users roles(RBAC), online editors(logs, statistical formulas, projects rules and etc) for more 10k clients.
-    Finalization coding script of internal protection algorithm authorization and validation.
-    Programming API service for quality control assessment. Interact with user interface(AJAX) with fast load up JSON data, audio files and extract large data in excel.
-    API data exchange integrated with data parse, merge, view in table linkage, send emails. Support more 100 source, near 1000 onlineusers, more 10 servers.
-    Codes cross-browsers users interfaces in project quality control assessment, using Javascript, jQuery, JSON, Bootstrap.
-    Developed JavaScript audio player with individual custom design, hardware acceleration, deceleration and the order to play audio files.
-
-  Designed and developed the company website (http://www.rozumsoft.com/).
-
-    Implemented 3 domain zones(ru/by/com) algorithm.
-    Codes contents editor for 3 languages
-    Fixed and support custom seo map logic.
-
-  Developed scripts products to callcenters operators
-
-    Development of сomplex reports and statistical summaries by Cisco data telephony.
-    Redesigned and reimplemented projects using MVC approach and strong OOP design
-    Designed and conversion of scripts database, extensive SQL query optimization.
-
-Real Estate Agency Assistant heals LLC, Full stack PHP Developer
-Belarus, Minsk, http://www.a-h.by; May 2011 - January 2012
-
-    Dynamic website design and programming using PHP, MySQL, HTML, CSS. Setup and administration of web servers and server software.
-    Business consulting of securing/ planning project.
-    Development to online marketing, search engine placement and promotion (http://www.mogu.by; http://www.a-h.by).
-
-EDUCATION
-
-    Belarusian University of Informatics and Radioelectronics,
-    BS in Modeling and computer-aided design of radioelectronics devices.
-
-Profiles
-https://github.com/aldb
-https://www.linkedin.com/pub/alex-dubinchyk/a0/54b/760/en
-
-skype: cool-skype-id
