--- storage/inputresume/rohit.txt
@@ -1,25 +1,0 @@
-Rohit
-(555) 888-0000
-gaurav@gmail.com
-
-OBJECTIVE:
-An MBA with 5 years of experience developing and managing marketing campaigns and specialized working knowledge of Google Analytics and AdWords, seeks the role of Social Media Marketing Manager with bnm Inc. to implement successful digital marketing campaigns and provide exceptional thought leadership.
-
-SUMMARY OF SKILLS:
-Three years of specialized experience in python , ML , AI ,
-Managed social media accounts for four businesses with yearly sales over $500,000
-Certified Digital Marketer 2016, Udemy
-Received Employee of the Year Award at bnm Company 2018 and 2019
-
-WORK EXPERIENCE:
-full stack web developer, ABC Company, Nashville, TN, 2017-2020
-Built a 15k Facebook following and 20k Instagram following for new startup, increasing leads by 160%
-Developed compelling CTAs for website and social media copy, producing a 60% increase in sales
-Formulated Instagram strategy and collaborated with influencers in our market to promote top-selling products, resulting in $200,000 net sales in three months
-Web developer, 123 Company, Austin, TX, 2014-2017
-
-Education:
-University of Michigan,2006-2010
-
-REFERENCES:
-Available upon request
