--- storage/output/emily_parsed.json
@@ -1,23 +1,0 @@
-{
-    "name": "e m i l y",
-    "email": "johndoe@email.com",
-    "phone": "1 2 3 4 5 6 7 8 9 0",
-    "objective": "Objective",
-    "summary": "",
-    "technology": "",
-    "skills": "Skills",
-    "experience": " Experienced in working on Agile teams\nWork Experience\n Gained experience in using Git and version control tools",
-    "education": "Education",
-    "languages": "",
-    "courses": "",
-    "projects": "",
-    "links": "",
-    "contacts": "",
-    "positions": "To obtain a position as a software developer",
-    "profiles": "",
-    "awards": "",
-    "honors": "",
-    "additional": "",
-    "certification": "",
-    "interests": ""
-}
