--- storage/output/yash_parsed.json
@@ -1,23 +1,0 @@
-{
-    "name": "y a s h",
-    "email": "john.smith@email.com",
-    "phone": "1 2 3 4 5 6 7 8 9 0",
-    "objective": "",
-    "summary": "",
-    "technology": "",
-    "skills": "Summary\nHighly motivated software engineer with 5 years of experience developing and implementing cuttingedge software solutions Proficient in Python  Java  ML and SQL  Skilled in data analysis and visualization Excellent problemsolving and communication skills",
-    "experience": "Experience\nSoftware Engineer ABC Company\nJune 2018  Present",
-    "education": "Education\nBachelor of Science in Computer Science Anytown University\nGraduated May 2016\nGPA 3950\n",
-    "languages": "",
-    "courses": "",
-    "projects": "",
-    "links": "",
-    "contacts": "",
-    "positions": "",
-    "profiles": "",
-    "awards": "",
-    "honors": "",
-    "additional": "",
-    "certification": "",
-    "interests": ""
-}
